feat: Update ToggleMenuFlyoutItem to winui3/release/1.5.4
